+++ packages/react-ui/components/Loader/Loader.tsx
@@ -10,6 +10,8 @@ import { Theme } from '../../lib/theming/Theme';
 import { ZIndex } from '../../internal/ZIndex';
 import { CommonWrapper, CommonProps } from '../../internal/CommonWrapper';
 import { cx } from '../../lib/theming/Emotion';
+import { isTestEnv } from '../../lib/currentEnvironment';
+import { TaskWithDelayAndMinimalDuration } from '../../lib/taskWithDelayAndMinimalDuration';
 import { getTabbableElements } from '../../lib/dom/tabbableHelpers';
 
 import { styles } from './Loader.styles';
@@ -28,10 +30,22 @@ export interface LoaderProps extends CommonProps {
   component?: React.ReactNode;
   className?: string;
   type?: 'mini' | 'normal' | 'big';
+  /**
+   * Время в миллисекундах для показа вуали без спиннера.
+   * @default 300
+   */
+  delayBeforeSpinnerShow: number;
+  /**
+   * Минимальное время в миллисекундах для показа спиннера
+   * @default 1000
+   */
+  minimalDelayBeforeSpinnerHide: number;
 }
 
 export interface LoaderState {
   isStickySpinner: boolean;
+  isSpinnerVisible: boolean;
+  isLoaderActive: boolean;
   spinnerStyle?: object;
 }
 
@@ -44,6 +58,8 @@ export class Loader extends React.Component<LoaderProps, LoaderState> {
   public static defaultProps: Partial<LoaderProps> = {
     type: Spinner.Types.normal,
     active: false,
+    delayBeforeSpinnerShow: isTestEnv ? 0 : 300,
+    minimalDelayBeforeSpinnerHide: isTestEnv ? 0 : 1000,
   };
 
   public static propTypes = {
@@ -74,6 +90,16 @@ export class Loader extends React.Component<LoaderProps, LoaderState> {
      * Spinner.types - все доступные типы
      */
     type: PropTypes.oneOf(Object.keys(Spinner.Types)),
+    /**
+     * Время в миллисекундах для показа вуали без спиннера.
+     * @default 300
+     */
+    delayBeforeSpinnerShow: PropTypes.number,
+    /**
+     * Минимальное время в миллисекундах для показа спиннера
+     * @default 1000
+     */
+    minimalDelayBeforeSpinnerHide: PropTypes.number,
   };
 
   private theme!: Theme;
@@ -81,6 +107,7 @@ export class Loader extends React.Component<LoaderProps, LoaderState> {
   private childrenContainerNode: Nullable<HTMLDivElement>;
   private spinnerNode: Nullable<HTMLDivElement>;
   private layoutEvents: Nullable<{ remove: () => void }>;
+  private spinnerTask: TaskWithDelayAndMinimalDuration;
   private childrenObserver: Nullable<MutationObserver>;
 
   constructor(props: LoaderProps) {
@@ -93,11 +120,21 @@ export class Loader extends React.Component<LoaderProps, LoaderState> {
 
     this.state = {
       isStickySpinner: false,
+      isSpinnerVisible: false,
+      isLoaderActive: false,
     };
+
+    this.spinnerTask = new TaskWithDelayAndMinimalDuration({
+      delayBeforeTaskStart: this.props.delayBeforeSpinnerShow,
+      durationOfTask: this.props.minimalDelayBeforeSpinnerHide,
+      taskStartCallback: () => this.setState({ isSpinnerVisible: true }),
+      taskStopCallback: () => this.setState({ isSpinnerVisible: false }),
+    });
   }
 
   public componentDidMount() {
     this.checkSpinnerPosition();
+    this.props.active && this.spinnerTask.start();
     this.layoutEvents = LayoutEvents.addListener(debounce(this.checkSpinnerPosition, 10));
 
     if (this.props.active) {
@@ -105,17 +142,34 @@ export class Loader extends React.Component<LoaderProps, LoaderState> {
     }
   }
 
-  public componentDidUpdate(prevProps: Readonly<LoaderProps>) {
-    const { component, active } = this.props;
+  public componentDidUpdate(prevProps: Readonly<LoaderProps>, prevState: Readonly<LoaderState>) {
+    const { component, active, delayBeforeSpinnerShow, minimalDelayBeforeSpinnerHide } = this.props;
+    const { isLoaderActive } = this.state;
 
     if ((active && !prevProps.active) || prevProps.component !== component) {
       this.checkSpinnerPosition();
     }
 
-    if (active) {
-      this.disableChildrenFocus();
-    } else {
-      this.enableChildrenFocus();
+    if (
+      delayBeforeSpinnerShow !== prevProps.delayBeforeSpinnerShow ||
+      minimalDelayBeforeSpinnerHide !== prevProps.minimalDelayBeforeSpinnerHide
+    ) {
+      this.spinnerTask.update({
+        delayBeforeTaskStart: delayBeforeSpinnerShow,
+        durationOfTask: minimalDelayBeforeSpinnerHide,
+      });
+    }
+
+    if (active !== prevProps.active) {
+      active ? this.spinnerTask.start() : this.spinnerTask.stop();
+    }
+
+    if (isLoaderActive !== prevState.isLoaderActive) {
+      if (isLoaderActive) {
+        this.disableChildrenFocus();
+      } else {
+        this.enableChildrenFocus();
+      }
     }
   }
 
@@ -124,6 +178,22 @@ export class Loader extends React.Component<LoaderProps, LoaderState> {
     if (this.layoutEvents) {
       this.layoutEvents.remove();
     }
+    this.spinnerTask.clearTask();
+  }
+
+  public static getDerivedStateFromProps(props: LoaderProps, state: LoaderState): Partial<LoaderState> {
+    if (props.active && !state.isLoaderActive) {
+      return {
+        isLoaderActive: true,
+      };
+    }
+    if (state.isLoaderActive && !(props.active || state.isSpinnerVisible)) {
+      return {
+        isLoaderActive: false,
+      };
+    }
+
+    return state;
   }
 
   public render() {
@@ -138,29 +208,30 @@ export class Loader extends React.Component<LoaderProps, LoaderState> {
   }
 
   private renderMain() {
-    const { active, type, caption, component } = this.props;
+    const { type, caption, component } = this.props;
+    const { isLoaderActive } = this.state;
 
     return (
       <CommonWrapper {...this.props}>
-        <div className={styles.loader()}>
+        <div className={styles.loader()} data-tid={isLoaderActive ? 'Loader__Veil' : ''}>
           <ZIndex
             priority={'Loader'}
-            applyZIndex={this.props.active}
-            coverChildren={this.props.active}
+            applyZIndex={isLoaderActive}
+            coverChildren={isLoaderActive}
             style={{ height: '100%' }}
             wrapperRef={this.childrenRef}
           >
             {this.props.children}
           </ZIndex>
-          {active && (
+          {isLoaderActive && (
             <ZIndex
               wrapperRef={this.spinnerRef}
               priority={'Loader'}
               className={cx({
-                [styles.active(this.theme)]: active,
+                [styles.active(this.theme)]: isLoaderActive,
               })}
             >
-              {this.renderSpinner(type, caption, component)}
+              {this.state.isSpinnerVisible && this.renderSpinner(type, caption, component)}
             </ZIndex>
           )}
         </div>
@@ -179,6 +250,7 @@ export class Loader extends React.Component<LoaderProps, LoaderState> {
   private renderSpinner(type?: 'mini' | 'normal' | 'big', caption?: React.ReactNode, component?: React.ReactNode) {
     return (
       <span
+        data-tid={'Loader__Spinner'}
         className={cx(styles.spinnerContainer(), { [styles.spinnerContainerSticky()]: this.state.isStickySpinner })}
         style={this.state.spinnerStyle}
       >