+++ packages/react-ui/components/Toggle/Toggle.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import warning from 'warning';
 
-import { tabListener } from '../../lib/events/tabListener';
+import { keyListener } from '../../lib/events/keyListener';
 import { ThemeContext } from '../../lib/theming/ThemeContext';
 import { Theme } from '../../lib/theming/Theme';
 import { CommonProps, CommonWrapper } from '../../internal/CommonWrapper';
@@ -77,7 +77,7 @@ export class Toggle extends React.Component<ToggleProps, ToggleState> {
 
   public componentDidMount() {
     if (this.props.autoFocus) {
-      tabListener.isTabPressed = true;
+      keyListener.isTabPressed = true;
       this.focus();
     }
   }
@@ -87,7 +87,7 @@ export class Toggle extends React.Component<ToggleProps, ToggleState> {
    */
   public focus = () => {
     if (this.input) {
-      tabListener.isTabPressed = true;
+      keyListener.isTabPressed = true;
       this.input.focus();
     }
   };
@@ -111,6 +111,7 @@ export class Toggle extends React.Component<ToggleProps, ToggleState> {
     const containerClassNames = cx(styles.container(this.theme), {
       [styles.containerDisabled(this.theme)]: !!disabled,
       [globalClasses.container]: true,
+      [globalClasses.containerDisabled]: !!disabled,
       [globalClasses.containerLoading]: loading,
     });
 
@@ -137,7 +138,7 @@ export class Toggle extends React.Component<ToggleProps, ToggleState> {
               [styles.focused(this.theme)]: !disabled && !!this.state.focusByTab,
             })}
           >
-            <span className={cx(styles.wrapper(this.theme), { [styles.wrapperDisabled()]: disabled })}>
+            <span className={cx(styles.wrapper(this.theme))}>
               <input
                 type="checkbox"
                 checked={checked}
@@ -151,7 +152,7 @@ export class Toggle extends React.Component<ToggleProps, ToggleState> {
               <div
                 className={containerClassNames}
                 style={
-                  checked && color
+                  checked && color && !disabled
                     ? {
                         backgroundColor: color,
                         boxShadow: `inset 0 0 0 1px ${color}`,
@@ -163,7 +164,14 @@ export class Toggle extends React.Component<ToggleProps, ToggleState> {
                   className={cx(styles.activeBackground(), globalClasses.background, {
                     [styles.activeBackgroundLoading(this.theme)]: loading,
                   })}
-                  style={checked && color ? { backgroundColor: color } : undefined}
+                  style={
+                    checked && color && !disabled
+                      ? {
+                          backgroundColor: color,
+                          boxShadow: `inset 0 0 0 1px ${color}`,
+                        }
+                      : undefined
+                  }
                 />
               </div>
               <div
@@ -204,7 +212,7 @@ export class Toggle extends React.Component<ToggleProps, ToggleState> {
       this.props.onFocus(event);
     }
 
-    if (tabListener.isTabPressed) {
+    if (keyListener.isTabPressed) {
       this.setState({ focusByTab: true });
     }
   };