+++ packages/react-ui/internal/themes/FlatTheme.ts
@@ -127,9 +127,9 @@ export class FlatTheme extends (class {} as typeof DefaultThemeInternal) {
   public static selectFontSizeMedium = '16px';
   //#endregion
   //#region Checkbox
-  public static checkboxBg = 'none';
+  public static checkboxBg = '#fff';
   public static get checkboxShadowDisabled() {
-    return `0 0 0 ${this.checkboxBorderWidth} #f2f2f2`;
+    return `0 0 0 ${this.checkboxBorderWidth} ${this.borderColorDisabled}`;
   }
   public static get checkboxShadowActive() {
     return `0 0 0 ${this.checkboxBorderWidth} #c3c3c3`;
@@ -163,9 +163,8 @@ export class FlatTheme extends (class {} as typeof DefaultThemeInternal) {
   //#region Input
   public static inputShadow = 'none';
   public static inputBorderTopColor = 'rgba(0, 0, 0, 0.15)';
-  public static inputDisabledBg = 'rgba(0, 0, 0, 0.05)';
   public static get inputDisabledBorderColor() {
-    return this.inputDisabledBg;
+    return this.borderColorDisabled;
   }
   //#endregion
   //#region Toggle
@@ -188,6 +187,7 @@ export class FlatTheme extends (class {} as typeof DefaultThemeInternal) {
   public static radioSizeAfter = '20px';
   public static radioVerticalAlign = 'top';
   public static radioBgImage = 'none';
+  public static radioBgColor = '#fff';
   public static radioBoxShadow = 'none';
   public static get radioBorder() {
     return `${this.radioBorderWidth} solid rgba(0, 0, 0, 0.15)`;
@@ -195,7 +195,9 @@ export class FlatTheme extends (class {} as typeof DefaultThemeInternal) {
   public static radioHoverShadow = 'none';
   public static radioActiveShadow = 'none';
   public static radioCheckedBulletColor = '#fff';
-  public static radioDisabledShadow = 'none';
+  public static get radioDisabledShadow() {
+    return `0 0 0 ${this.radioBorderWidth} ${this.borderColorDisabled}`;
+  }
   public static radioLabelDisplay = 'inline-block';
   public static get radioHoverBg() {
     return this.checkboxHoverBg;
@@ -213,9 +215,17 @@ export class FlatTheme extends (class {} as typeof DefaultThemeInternal) {
   public static get radioCheckedHoverBgColor() {
     return ColorFunctions.darken(this.checkboxCheckedBg, '5%');
   }
-  public static switcherButtonDisabledBorderColor: 'rgba(0, 0, 0, 0.15)';
+  public static get switcherButtonDisabledBorderColor() {
+    return this.borderColorDisabled;
+  }
 
   public static switcherButtonCheckedDisabledShadow: 'none';
+  //#endregion
+  //#region Token
+  public static tokenDisabledBg = 'rgba(0, 0, 0, 0.05)';
+  public static get tokenShadowDisabled() {
+    return `0 0 0 ${this.tokenBorderWidth} ${this.borderColorDisabled}`;
+  }
   //#endregion
 }