+++ packages/react-ui/lib/events/tabListener.ts
@@ -1,15 +1,2 @@
-import { isBrowser } from '../client';
-
-import { isKeyTab } from './keyboard/identifiers';
-
-class TabListener {
-  public isTabPressed = false;
-  constructor() {
-    if (isBrowser) {
-      window.addEventListener('keydown', (e) => (this.isTabPressed = isKeyTab(e)));
-      window.addEventListener('mousedown', () => (this.isTabPressed = false));
-    }
-  }
-}
-
-export const tabListener = new TabListener();
+// deprecated, use keyListener
+export { keyListener as tabListener } from './keyListener';