+++ packages/react-ui/CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.7.0](https://github.com/skbkontur/retail-ui/compare/@skbkontur/react-ui@3.6.0...@skbkontur/react-ui@3.7.0) (2021-09-16)
+
+
+### Bug Fixes
+
+* **checkbox:** add `!important` for focus styles ([#2528](https://github.com/skbkontur/retail-ui/issues/2528)) ([35c6a69](https://github.com/skbkontur/retail-ui/commit/35c6a6905a6722ca0b99cb86aadee4731effc0ff)), closes [#2517](https://github.com/skbkontur/retail-ui/issues/2517)
+* **Loader:** fix focus disabling ([fcc5e1e](https://github.com/skbkontur/retail-ui/commit/fcc5e1e42dcacffcb5d0abcb79998c3e0a107e2e))
+* **radio:** remove focus styles on mouse click ([876a1b8](https://github.com/skbkontur/retail-ui/commit/876a1b87ce5b0a16fe212b5a74680be7e20e1b3b))
+* **theme:** fix disabled style of 4 controls* ([#2524](https://github.com/skbkontur/retail-ui/issues/2524)) ([6783532](https://github.com/skbkontur/retail-ui/commit/6783532cedde0c09ff1bcaaddc712c41cf9dbc9e))
+* **theme:** remove color variables from 8px theme ([c86ed5d](https://github.com/skbkontur/retail-ui/commit/c86ed5dd754affd93e1727305eb1f34189a5d98d))
+* **theme:** to correct disabled style of Input ([#2521](https://github.com/skbkontur/retail-ui/issues/2521)) ([d91fedd](https://github.com/skbkontur/retail-ui/commit/d91feddf0fbccc9d0129552c465e7fc29f4dbdff))
+
+
+### Features
+
+* **ScrollContainer:** horizontal scroll ([#2473](https://github.com/skbkontur/retail-ui/issues/2473)) ([1df7613](https://github.com/skbkontur/retail-ui/commit/1df7613b9d16c490bb110c3f872ee03152f5f45b)), closes [#1931](https://github.com/skbkontur/retail-ui/issues/1931)
+
+
+
+
+
+# [3.6.0](https://github.com/skbkontur/retail-ui/compare/@skbkontur/react-ui@3.5.0...@skbkontur/react-ui@3.6.0) (2021-08-26)
+
+
+### Bug Fixes
+
+* **FlatTheme:** set white bg for Radio, Checkbox and Toggle ([#2486](https://github.com/skbkontur/retail-ui/issues/2486)) ([07fac23](https://github.com/skbkontur/retail-ui/commit/07fac23f0089704db6a71e9561f0c5761df2d7dd))
+* **Loader:** disable focus children events while page is loading ([#2494](https://github.com/skbkontur/retail-ui/issues/2494)) ([1d46f96](https://github.com/skbkontur/retail-ui/commit/1d46f9667d5e1d3cfb55a131687210c4ca0948a8)), closes [#1732](https://github.com/skbkontur/retail-ui/issues/1732)
+
+
+### Features
+
+* **Button:** add `btnWarningSecondary` theme var and * ([#2512](https://github.com/skbkontur/retail-ui/issues/2512)) ([8a75266](https://github.com/skbkontur/retail-ui/commit/8a75266d65b73f159461313b7ec35f014ce02b37))
+* **Loader:** add props for show loader without spinner ([#2445](https://github.com/skbkontur/retail-ui/issues/2445)) ([2ffa70b](https://github.com/skbkontur/retail-ui/commit/2ffa70b165af3d31f2e6da06214d91b07c031563)), closes [#2266](https://github.com/skbkontur/retail-ui/issues/2266)
+
+
+
+
+
 # [3.5.0](https://github.com/skbkontur/retail-ui/compare/@skbkontur/react-ui@3.4.0...@skbkontur/react-ui@3.5.0) (2021-08-17)